# map: remove malloc and memcpy from get