Remove the copying of the output value. Much better performance:

Before:
```
132318         42.493ms            321ns map_get3 
```

After:
```
132296         33.781ms            255ns map_get3 
```